+++ TrainingMesh/Utils.cpp
@@ -15,8 +15,8 @@ bool Utils::onSegment(float x1, float y1, float x2, float y2, float x3, float y3
 // 2 --> Counterclockwise
 int Utils::orientation(float x1, float y1, float x2, float y2, float x3, float y3)
 {
-	int val = (y2 - y1) * (x3 - x2) -
-		(x2 - x1) * (y3 - y2);
+	int val = (int)((y2 - y1) * (x3 - x2) -
+		(x2 - x1) * (y3 - y2));
 
 	if (val == 0) return 0;  // colinear
